refactor: move runner group check from zero/org to infra/run
Move the isOfficialRunnerGroup function from zero/org/org-service.ts
to infra/run/runner-group.ts to eliminate the infra→zero reverse
dependency. Update all 4 consumers (runner-executor + 3 API routes).

Closes #7789
