+++ turbo/apps/web/src/lib/infra/run/executors/runner-executor.ts
@@ -11,7 +11,7 @@ import {
   type RunContextSnapshot,
 } from "../../../shared/axiom/client";
 import { encryptSecretsMap } from "../../../shared/crypto/secrets-encryption";
-import { isOfficialRunnerGroup } from "../../../zero/org/org-service";
+import { isOfficialRunnerGroup } from "../runner-group";
 import { forbidden } from "../../../errors";
 import { publishJobNotification } from "../../../shared/realtime/client";
 import { logger } from "../../../logger";
